+++ LookupAnything/docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * Fixed machine output recipes no longer shown.
+* Fixed cooking recipes always shown as never cooked.
 * Improved translations. Thanks to MakinDay (updated Italian), mc-kaishixiaxue (updated Chinese), and Shi974 (updated French)!
 
 ## 1.41.2